% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_tempdir}
\alias{add_tempdir}
\title{Add the output of \code{\link{tempdir}} to a makelist}
\usage{
add_tempdir(ml)
}
\arguments{
\item{ml}{The makelist.}
}
\value{
The makelist.
}
\description{
You do not want to litter your working directory? Use \R's temporary
directory.
}
\note{
This is mainly meant to run examples without touching your disk. You
should not bother.
}
\examples{
str(add_tempdir(provide_make_list("minimal")))
}
