% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_archive_as_cran}
\alias{check_archive_as_cran}
\title{A Convenience Wrapper to \code{\link{check_archive}}}
\usage{
check_archive_as_cran(path)
}
\arguments{
\item{path}{Path to the package archive.}
}
\description{
A Convenience Wrapper to \code{\link{check_archive}}
}
