% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakemake-package.R
\docType{package}
\name{fakemake-package}
\alias{fakemake-package}
\title{Mock the Unix Make Utility}
\description{
Use R as a minimal build system. This might come in handy if you are
developing R packages and can not use a proper build system.
}
\details{
You will probably want to start reading about
\code{\link[fakemake:make]{make}} or reading the \cr
\code{vignette("An_Introduction_to_fakemake", package = "fakemake")}.
}
\keyword{package}
