% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makelist2igraph}
\alias{makelist2igraph}
\title{Parse a \code{Makelist} and Convert it Into an \code{Igraph}}
\usage{
makelist2igraph(make_list)
}
\arguments{
\item{make_list}{The \code{makelist}.}
}
\value{
An \pkg{igraph} representation of the \code{makelist}.
run.
}
\description{
Parse a \code{Makelist} and Convert it Into an \code{Igraph}
}
\examples{
str(ml <- provide_make_list("package"))
plot(makelist2igraph(ml))
}
