% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelists.R
\name{remove_target}
\alias{remove_target}
\title{Remove a Target From a Makelist}
\usage{
remove_target(makelist, target)
}
\arguments{
\item{makelist}{A list for
\code{\link[fakemake:make]{make}}.}

\item{target}{The target to remove from \code{makelist}.}
}
\value{
A list for
\code{\link[fakemake:make]{make}}.
}
\description{
Remove a target and all its appearances as
other targets' dependencies from a \code{makelist}.
}
\seealso{
Other functions to manipulate makelists: 
\code{\link{add_target}()},
\code{\link{get_target}()}
}
\concept{functions to manipulate makelists}
