% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{touch}
\alias{touch}
\title{Mock the Unix \command{touch} utility}
\usage{
touch(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[fritools:touch]{fritools::touch}}.}
}
\value{
The return value of \code{\link[fritools:touch]{fritools::touch}}.
}
\description{
See \code{\link[fritools:touch]{fritools::touch}}.
}
\keyword{internal}
