% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{visualize}
\alias{visualize}
\title{Visualize a Makelist}
\usage{
visualize(make_list, root = NULL)
}
\arguments{
\item{make_list}{The \code{makelist}.}

\item{root}{The root of a tree.}
}
\value{
Invisibly an \pkg{igraph} representation of the \code{makelist}.
}
\description{
Parse a \code{makelist}, convert it into an \code{igraph} and plot it.
}
\examples{
str(ml <- provide_make_list())
visualize(ml)
visualize(ml, root = "all.Rout")
}
