% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_makefile}
\alias{write_makefile}
\title{Write a Makelist to File}
\usage{
write_makefile(make_list, path, Rbin = "Rscript-devel")
}
\arguments{
\item{make_list}{The list to write to file.}

\item{path}{The path to the file.}

\item{Rbin}{The R binary to use in the Makefile.}
}
\value{
See 
\code{\link[MakefileR:write_makefile]{MakefileR::write_makefile}}.
}
\description{
Write a Makelist to File
}
\examples{
make_file <- file.path(tempdir(), "Makefile")
write_makefile(provide_make_list(), path = make_file)
file.show(make_file, pager = "cat")
}
