% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakemake-package.R
\docType{package}
\name{fakemake-package}
\alias{fakemake-package}
\title{Mock the Unix Make Utility}
\description{
A description is a paragraph consisting of one or more sentences.
You may add another paragraph for a 'Details' section.
}
\keyword{package}
