% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make}
\alias{make}
\title{Mock the Unix Make Utility}
\usage{
make(target, make_list)
}
\arguments{
\item{target}{The make target.}

\item{make_list}{The makelist (a listed version of a Makefile).}
}
\value{
\code{\link[base:NULL]{base::NULL}}.
}
\description{
Mock the Unix Make Utility
}
\examples{
str(make_list <- provide_make_list())
make("all.Rout", make_list)
}
