% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{provide_make_list}
\alias{provide_make_list}
\title{Load an Example \code{Makelist} Provided by \pkg{fakemake}.}
\usage{
provide_make_list(type = "minimal")
}
\arguments{
\item{type}{The type of \code{makelist}.}
}
\value{
A \code{makelist}.
}
\description{
Load an Example \code{Makelist} Provided by \pkg{fakemake}.
}
\examples{
str(provide_make_list("minimal"))
}
