% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_makefile}
\alias{read_makefile}
\title{Read a Makefile into a Makelist}
\usage{
read_makefile(path)
}
\arguments{
\item{path}{The path to the file.}
}
\value{
The makelist.
}
\description{
Read a Makefile into a Makelist
}
\note{
This function will not read arbitrary Makefiles, just those
created via \code{\link{write_makefile}}! If you modify such a Makefile
make sure you only add simple rules like the ones you see in that file.
}
\examples{
make_file <- file.path(tempdir(), "Makefile")
write_makefile(provide_make_list(), path = make_file)
str(make_list <- read_makefile(path = make_file))
}
