% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{touch}
\alias{touch}
\title{Mock the Unix \command{touch} utility}
\usage{
touch(path)
}
\arguments{
\item{path}{Path to the file to be touched}
}
\value{
The return value of \code{\link{file.copy}}.
}
\description{
See \code{\link[fritools:touch]{fritools::touch}}.
}
\keyword{internal}
