% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{visualize}
\alias{visualize}
\title{Parse a \code{Makelist}, Convert it Into an \code{Igraph} and Plot it}
\usage{
visualize(make_list, root = NULL)
}
\arguments{
\item{make_list}{The \code{makelist}.}

\item{root}{The root of a tree.}
}
\value{
Invisibly an \pkg{igraph} representation of the \code{makelist}.
}
\description{
Parse a \code{Makelist}, Convert it Into an \code{Igraph} and Plot it
}
\examples{
str(ml <- provide_make_list("package"))
visualize(ml)
visualize(ml, root = "log/check.Rout")
}
